requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
httpx==0.25.0
python-multipart==0.0.6
python-jose==3.3.0
bcrypt==4.0.1
PyJWT==2.8.0
pydantic==2.3.0
PyPDF2==3.0.1
python-docx==0.8.11